+++ utilities/database/orm/__init__.py
@@ -69,3 +69,4 @@ class SlackSettings:
     automated_preblast_option: Optional[str] = None
     automated_preblast_hour_cst: Optional[int] = None
     scheduled_preblast_hour_cst: Optional[int] = None
+    hc_announce_option: Optional[str] = None